Adding a rough sketch of a test
