+++ wdfn-server/waterdata/tests/test_views.py
@@ -194,12 +194,6 @@ class TestTimeSeriesComponentView:
         assert text.count('class="wdfn-component" data-component="hydrograph"') == 1, 'Component expected'
 
 
-class TestCooperatorLookup:
-    @mock.patch('waterdata.views.requests.get()')
-    def mock_site_call(self):
-        assert 0 == 0
-
-
 MOCK_SITE_LIST_1 = (
     '#\n#\n# US Geological Survey\n# retrieved: 2018-01-02 09:31:20 -05:00\t(caas01)\n#\n# '
     'The Site File stores location and general information about groundwater,\n# surface '
@@ -259,7 +253,3 @@ MOCK_SITE_LIST_2 = (
 )
 
 
-MOCK_COOPERATOR_LOOKUP_DATA = '{"Customers":[{"Name":"New Jersey Department of Environmental Protection",' \
-                  '"URL":"http://www.nj.gov/dep/ec/","IconURL":"http://water.usgs.gov/customer/icons/1275.gif"},' \
-                  '{"Name":"USGS - National Groundwater Monitoring Network",' \
-                  '"URL":"http://www.usgs.gov/","IconURL":"http://water.usgs.gov/customer/icons/9326.gif"}]}'
