+++ wdfn-server/waterdata/views.py
@@ -58,8 +58,9 @@ def monitoring_location(site_no):
         station_record = data_list[0]
 
         # get the cooperator data from service
-        if app.config['COOPERATOR_LOOKUP_ENABLED']:  # feature toggle for stopping use on production
-            cooperator_lookup_data = execute_lookup_request(app.config['URL_ROOT_COOPERATOR_LOOKUP'], site_no, app.config['URL_PARAMS_COOPERATOR_LOOKUP'])
+        if app.config['COOPERATOR_LOOKUP_ENABLED']:  # feature toggle; remove 'if/else' when new lookup service is implemented
+            cooperator_lookup_data = execute_lookup_request(app.config['URL_ROOT_COOPERATOR_LOOKUP'],
+                                                            site_no, app.config['URL_PARAMS_COOPERATOR_LOOKUP'])
         else:
             cooperator_lookup_data = None
